+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Increased `dartfmt` timeout to 5 minutes.
 
+* Store basic stats in the analysis summary.
+
 ## 0.12.2
 
 * Bugfix: pre-v1 release should be pre-v0.1 instead.